upload the 1.0 version of the code
This code was written during my MSC thesis, I decided to shere it
everybody. You can read about the details, the goal of being this code
in the Readme file.

This is still a spagetti code, it requires some improuvement. There no
test to this code, I am working on it...